refactor steps component
