+++ cmd/component/steps/entry.go
@@ -1,0 +1,15 @@
+package steps
+
+type Entry struct {
+	Text   string
+	Status int
+	Err    error
+}
+
+func NewStep(text string) Entry {
+	return Entry{
+		Text:   text,
+		Status: INCOMPLETE,
+		Err:    nil,
+	}
+}
